Move pubspec suggestors
