+++ lib/src/executables/react17_dependency_override_update.dart
@@ -15,7 +15,7 @@
 import 'dart:io';
 
 import 'package:codemod/codemod.dart';
-import 'package:over_react_codemod/src/react16_suggestors/dependency_override_updater.dart';
+import 'package:over_react_codemod/src/pubspec_suggestors/dependency_override_updater.dart';
 import 'package:over_react_codemod/src/util.dart';
 
 const _changesRequiredOutput = """
