+++ lib/src/pubspec_suggestors/pubspec_upgrader.dart
